+++ L4D2AddonAssistant.GUI/IAppWindowManager.cs
@@ -1,9 +1,15 @@
-﻿using System;
+﻿using L4D2AddonAssistant.ViewModels;
+using L4D2AddonAssistant.Views;
+using System;
 
 namespace L4D2AddonAssistant
 {
     public interface IAppWindowManager
     {
+        MainWindow? MainWindow { get; }
+
+        MainWindow CreateMainWindow(MainWindowViewModel viewModel);
+
         void OpenSettingsWindow();
 
         void OpenDownloadListWindow();