+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - fix: renaming bug
 - improve: change the deletion behavior from direct deletion to move to recycle bin (Windows only)
-- add: show in the file explorer
+- add: show in the file explorer (Windows only)
+- add: auto detect workshop item link in the clipboard
 # v0.3.0
 - add: creation time display
 - add: sort by creation time